requirements_python3131.txt
# Updated requirements.txt for Python 3.13.1 compatibility
# Core GUI and data processing
pandas>=2.0.0
pillow>=10.0.0
openpyxl>=3.1.0
tkinter>=0.1.0

# Build tools
pyinstaller>=6.0.0
setuptools>=68.0.0
wheel>=0.41.0

# Optional performance improvements for Python 3.13.1
numpy>=1.26.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development and testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# System compatibility
pywin32>=306; sys_platform == "win32"
pywin32-ctypes>=0.2.2; sys_platform == "win32"
